alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { FlatList, View } from 'react-native'
+import { FlatList, View, Alert, Text } from 'react-native'
 import { ListItem } from 'react-native-elements'
 import { connect } from 'react-redux'
 import { baseUrl } from '../shared/baseUrl'
@@ -29,10 +29,27 @@ class Favorites extends React.Component{
         const renderMenuItem = ({item, index}) => {
 
             const rightButton = [
-                {
-                    text: 'Delete',
-                    type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                { text: 'Delete', 
+                type: 'delete',
+                onPress: () => {
+                    Alert.alert(
+                        'Delete Favorite?',
+                        'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                        [
+                            { 
+                                text: 'Cancel', 
+                                onPress: () => console.log(item.name + 'Not Deleted'),
+                                style: ' cancel'
+                            },
+                            {
+                                text: 'OK',
+                                onPress: () => this.props.deleteFavorite(item.id)
+                            }
+                        ],
+                        { cancelable: false }
+                    );
+                    
+                }
                 }
             ]
